start moving worst report/testrunner integration test into its own module
